Move xpr-types to xpr to simplify diff
